# packages/admin-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shipeasy Admin API",
    "version": "1.0.0",
    "description": "REST API for managing feature gates, experiments, configs, universes, and killswitches in a Shipeasy project. Authenticate with an admin SDK key (`Authorization: Bearer sdk_admin_…`) and scope every request to a project via the `X-Project-Id` header.\n\nMint admin keys via `POST /api/admin/keys` with `type: \"admin\"`. Keys expire after 90 days; rotate with the `revoke` action.",
    "contact": {
      "name": "Shipeasy",
      "url": "https://shipeasy.ai"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://shipeasy.ai",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Next.js dev server"
    }
  ],
...
  "paths": {
    "/api/admin/gates": {
      "get": {
        "operationId": "listGates",
        "summary": "List feature gates",
        "description": "Returns a single page of gates ordered by `updated_at desc, id desc`. Use the `cursor` query parameter to paginate.\n\n**Use case:** Snapshot every gate in the project — for example to render an admin overview or to drive a CI check that asserts no gate is left at 100% in staging.",
        "tags": ["Gates"],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/PaginationLimit"
          },
          {
            "$ref": "#/components/parameters/PaginationCursor"
          }
        ],
        "responses": {
          "200": {
            "description": "List feature gates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListGatesResponse"
                },
                "example": {
                  "data": [
                    {
                      "id": "gat_01j7w7m9q4hxbf6npe6s9zr3vc",
                      "name": "checkout_v2",
                      "enabled": 1,
                      "rolloutPct": 5000,
                      "rules": [],
                      "updatedAt": "2026-05-09T16:01:22.000Z"
                    }
                  ],
                  "next_cursor": null
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          }
        }
      },
      "post": {
        "operationId": "createGate",
        "summary": "Create a feature gate",
        "description": "Creates a new gate, defaulting to enabled at the supplied rollout.\n\n**Use case:** Create the gate at 0% rollout from your release pipeline, then ramp it via PATCH once you've validated the rollout in production.",
        "tags": ["Gates"],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGateRequest"
              },
              "example": {
                "name": "checkout_v2",
                "rollout_pct": 0
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a feature gate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateGateResponse"
                },
                "example": {
                  "id": "gat_01j7w7m9q4hxbf6npe6s9zr3vc",
                  "name": "checkout_v2"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          },
          "code": {
            "type": "string",
            "description": "Stable machine code, when applicable."
          },
          "detail": {
            "type": "string",
            "description": "Extra context (stack frames, validation details)."
          }
        },
        "additionalProperties": false
      },
      "ListGatesResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Stable opaque gate id."
                },
                "name": {
                  "type": "string",
                  "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
                },
                "enabled": {
                  "anyOf": [
                    {
                      "type": "boolean"
                    },
                    {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1
                    }
                  ]
                },
                "rolloutPct": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 10000
                },
                "rules": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "attr": {
                        "type": "string",
                        "minLength": 1
                      },
                      "op": {
                        "type": "string",
                        "enum": [
                          "eq",
                          "neq",
                          "in",
                          "not_in",
                          "gt",
                          "gte",
                          "lt",
                          "lte",
                          "contains",
                          "regex"
                        ]
                      },
                      "value": {}
                    },
                    "required": ["attr", "op", "value"],
                    "additionalProperties": false
                  }
                },
                "salt": {
                  "type": "string"
                },
                "title": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "description": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "folder": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "groupName": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "ownerEmail": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "stack": {
                  "anyOf": [
                    {
                      "type": "array",
                      "items": {
                        "oneOf": [
                          {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "minLength": 1
                              },
                              "type": {
                                "type": "string",
                                "const": "condition"
                              },
                              "name": {
                                "type": "string"
                              },
                              "fromTemplate": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "pass": {
                                "type": "string",
                                "enum": ["all", "any"]
                              },
                              "rules": {
                                "default": [],
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "attr": {
                                      "type": "string",
                                      "minLength": 1
                                    },
                                    "op": {
                                      "type": "string",
                                      "enum": [
                                        "eq",
                                        "neq",
                                        "in",
                                        "not_in",
                                        "gt",
                                        "gte",
                                        "lt",
                                        "lte",
                                        "contains",
                                        "regex"
                                      ]
                                    },
                                    "value": {}
                                  },
                                  "required": ["attr", "op", "value"],
                                  "additionalProperties": false
                                }
                              },
                              "locked": {
                                "type": "boolean"
                              }
                            },
                            "required": ["id", "type", "rules"],
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "minLength": 1
                              },
                              "type": {
                                "type": "string",
                                "const": "rollout"
                              },
                              "name": {
                                "type": "string"
                              },
                              "fromTemplate": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "rolloutPct": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 10000
                              },
                              "bucketBy": {
                                "type": "string"
                              },
                              "salt": {
                                "type": "string"
                              },
                              "locked": {
                                "type": "boolean"
                              }
                            },
                            "required": ["id", "type", "rolloutPct"],
                            "additionalProperties": false
                          }
                        ]
                      }
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "updatedAt": {
                  "type": "string",
                  "description": "ISO-8601 timestamp of last mutation."
                }
              },
              "required": ["id", "name", "enabled", "rolloutPct", "updatedAt"],
              "additionalProperties": false
            }
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": ["data", "next_cursor"],
        "additionalProperties": false
      },
      "CreateGateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
          },
          "rollout_pct": {
            "default": 0,
            "type": "integer",
            "minimum": 0,
            "maximum": 10000
          },
          "rules": {
            "default": [],
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "attr": {
                  "type": "string",
                  "minLength": 1
                },
                "op": {
                  "type": "string",
                  "enum": [
                    "eq",
                    "neq",
                    "in",
                    "not_in",
                    "gt",
                    "gte",
                    "lt",
                    "lte",
                    "contains",
                    "regex"
                  ]
                },
                "value": {}
              },
              "required": ["attr", "op", "value"],
              "additionalProperties": false
            }
          },
          "salt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          }
        },
        "required": ["name", "rollout_pct", "rules"],
        "additionalProperties": false
      },
      "CreateGateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
          }
        },
        "required": ["id", "name"],
        "additionalProperties": false
      },
...
    },
    "parameters": {
      "ProjectId": {
        "name": "X-Project-Id",
        "in": "header",
        "required": true,
        "description": "Project the request operates on. Must match the project the SDK key belongs to.",
        "schema": {
          "type": "string"
        }
      },
      "PaginationLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Page size (1–500). Defaults to 100.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 500,
          "default": 100
        }
      },
      "PaginationCursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque cursor returned in the previous page's `next_cursor`. Omit for the first page.",
        "schema": {
          "type": "string"
        }
      }
    },
    "securitySchemes": {
      "bearerSdkKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "sdk_admin_*",
        "description": "Pass an admin SDK key as `Authorization: Bearer sdk_admin_…`. Mint via `POST /api/admin/keys` with `type: \"admin\"`."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request was malformed or failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid SDK key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "SDK key is valid but not allowed to act on this project.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource does not exist or is not visible to the caller.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "The mutation conflicts with current state (e.g. duplicate name, or a referenced row is still in use).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Zod schema validation rejected the request body.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerSdkKey": []
    }
  ]
}